AWS: Improve generic remote object description
The get_action binding method can now handle the exceptions
AWS sometimes throws to mean 404.

The describe_filters method can return None to indicate
that we have determined an object definitely can't exist.
For example, an AWS subnet. We know we don't have a vpc id.
And we need that to find the subnet. So the subnet can't
exist.