AWS IAM: Normalize certificate body and chains.
AWS will accept various variants of PEM encoding, but it will normalize them.
So if you get a PEM cert from a Windows box you might find the line length is
different and that line endings are different. If AWS normalizes the cert but
we don't then we can't ever match that cert. Post creation 'did it work' tests
will always fail.

Now we normalize both the variables from the user and from AWS so we know they
are in the same variant of the PEM format, allowing comparisons to succeed.